Module Runtime.Block

Represents a single task or unit of behavior. Blocks are Objective-C objects, which means they can be added to collections like NSArray or NSDictionary.

type t

val t : t Ctypes.structure Ctypes.typ

val isa : (Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr, t Ctypes.structure) Ctypes.field

val flags : (int, t Ctypes.structure) Ctypes.field

val reserved : (int, t Ctypes.structure) Ctypes.field

val invoke : (unit Ctypes_static.ptr, t Ctypes.structure) Ctypes.field

val descriptor : (Block_descriptor.t Ctypes.structure Ctypes_static.ptr, t Ctypes.structure) Ctypes.field

val size : int

val desc_ptr : Block_descriptor.t Ctypes.structure Ctypes.ptr

val block_is_global : int

val self : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr

val make' : ('a -> 'b) -> typ:('a -> 'b) Ctypes.fn -> unit Ctypes_static.ptr

val make : (Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr -> 'a) -> args:('a, 'b) Objc_type.tlist -> return:'b Objc_type.t -> unit Ctypes_static.ptr

Create a global block which encapsulates the code for execution at a later time.