Module Runtime.Class

val get_name : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> string

Returns the name of a class.

val get_superclass : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr

Returns the superclass of a class.

val add_protocol : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> bool

Adds a protocol to a class.

val conforms_to_protocol : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> bool

Returns a Boolean value that indicates whether a class conforms to a given protocol.

val copy_method_list : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Unsigned.uint Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr Ctypes_static.ptr

Describes the instance methods implemented by a class. outCount

On return, contains the length of the returned array. If outCount is NULL, the length is not returned.

val get_instance_method : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr

Returns a specified instance method for a given class.

val get_instance_variable : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> string -> Runtime__.Type_description.Types.objc_ivar Ctypes.structure Ctypes_static.ptr

Returns the Ivar for a specified instance variable of a given class.

val get_method_implementation : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr -> unit Ctypes_static.ptr

Returns the function pointer that would be called if a particular message were sent to an instance of a class.

val is_meta_class : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> bool

Returns a Boolean value that indicates whether a class object is a metaclass.

val copy_property_list : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Unsigned.uint Ctypes_static.ptr -> [ `Prop_attr ] Ctypes.structure Ctypes_static.ptr

Describes the properties declared by a class.

val add_property : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> string -> [ `Prop_attr ] Ctypes.structure Ctypes_static.ptr -> Unsigned.uint -> bool

Adds a property to a class.

val get_property : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> string -> [ `Prop_attr ] Ctypes.structure Ctypes_static.ptr

Returns a property with a given name of a given class.

val defined_method_roots : unit Ctypes.ptr list Stdlib.ref

val alignment_of_size : Unsigned.Size_t.t -> Unsigned.UInt8.t

val create_instance : ?extra_bytes:Unsigned.size_t -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr

val add_method : self: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> cmd: Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr -> typ:'a Ctypes.fn -> enc:string -> ?runtime_lock:bool -> ?thread_registration:bool -> (Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr -> 'a) -> bool

Adds a new method to a class with a given name and implementation.

val add_ivar : self: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> name:string -> size:Unsigned.size_t -> enc:string -> bool

Adds a new instance variable to a class.

val define : ?superclass: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> ?protocols: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr list -> ?ivars:Define.ivar_spec' list -> ?properties:Define.prop_spec' list -> ?methods:Define.method_spec' list -> ?class_methods:Define.method_spec' list -> string -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr

Defines a new class and registers it with the Objective-C runtime.