Module Runtime.Method

val get_name : Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr

Returns the name of a method.

val get_return_type : Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr -> string

Returns a string describing a method's return type.

val get_number_of_arguments : Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr -> Unsigned.uint

Returns the number of arguments accepted by a method.

val get_argument_type : Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr -> Unsigned.uint -> string

Returns a string describing a single parameter type of a method.

val get_type_encoding : Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr -> string

Returns a string describing a method's parameter and return types.

val invoke : typ:'a Ctypes.fn -> self: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_method Ctypes.structure Ctypes_static.ptr -> 'a

val define : cmd: Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes.ptr -> args:('a, 'b) Objc_type.tlist -> return:'b Objc_type.t -> ?runtime_lock:bool -> ?thread_registration:bool -> (Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes.ptr -> 'a) -> Define.method_spec'