Module Runtime.Objc_type

type _ t =

| Id : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr t

| Class : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr t

| Sel : Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes.ptr t

| Void : unit t

| Str : string t

| Char : char t

| Bool : bool t

| Int : int t

| Short : int t

| Long : Signed.long t

| LLong : Signed.llong t

| ULLong : Unsigned.ullong t

| Float : float t

| Double : float t

| Unknown : unit Ctypes.ptr t

| Ptr : 'a t -> 'a Ctypes.ptr t

| Arr : 'a t -> 'a Ctypes.carray t

| Struc : 'a t -> 'a Ctypes.structure t

| Union : 'a t -> 'a Ctypes.union t

| Imp : unit Ctypes.ptr t

| Enc : string t

| Proto : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr t

| Ivar : Runtime__.Type_description.Types.objc_ivar Ctypes.structure Ctypes.ptr t

Represents an Objective-C type.

type (_, _) tlist =

| [] : ('r, 'r) tlist

| :: : 'a t * ('b, 'r) tlist -> ('a -> 'b, 'r) tlist

Represents a tuple of Objective-C types.

val method_typ : args:('a, 'b) tlist -> 'b t -> 'a Ctypes.fn

Returns the ctypes fn typ corresponding to the Objective-C method type.

val value_typ : 'a t -> 'a Ctypes.typ

Returns the ctypes value typ corresponding to the Objective-C value type.

val encode_method : args:('a, 'b) tlist -> 'c t -> string

Returns the encoding of a method type in the Objective-C runtime.

val encode_value : 'a t -> string

Returns the encoding of a value type in the Objective-C runtime.

val noargs : ('a, 'a) tlist

The method accepts only the implicit self and cmd arguments.

val id : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr t

A reference to any Objective-C object regardless of its class.

val _Class : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr t

val _SEL : Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes.ptr t

val _IMP : unit Ctypes.ptr t

val _Enc : string t

val _Protocol : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr t

val _Ivar : Runtime__.Type_description.Types.objc_ivar Ctypes.structure Ctypes.ptr t

val void : unit t

val string : string t

val char : char t

val bool : bool t

val int : int t

val short : int t

val long : Signed.long t

val llong : Signed.llong t

val ullong : Unsigned.ullong t

val float : float t

val double : float t

val unknown : unit Ctypes.ptr t

val ptr : 'a t -> 'a Ctypes.ptr t

val array : 'a t -> 'a Ctypes.carray t

val struc : 'a t -> 'a Ctypes.structure t

val union : 'a t -> 'a Ctypes.union t