Module Runtime.Object

val get_class : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr

Returns the class of an object.

val get_ivar : self: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> ivar: Runtime__.Type_description.Types.objc_ivar Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr

Reads the value of an Id instance variable in an object.

val set_ivar : self: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> ivar: Runtime__.Type_description.Types.objc_ivar Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> unit

Sets the value of an instance variable in an object. object_setIvar is faster than object_setInstanceVariable if the Ivar for the instance variable is already known.

val get_instance_variable : self: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> name:string -> value_ptr:'a Ctypes_static.ptr -> typ:'a Ctypes.typ -> Runtime__.Type_description.Types.objc_ivar Ctypes.structure Ctypes_static.ptr

Obtains the value of an instance variable of a class instance.

val set_instance_variable : self: Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr -> name:string -> value:unit Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_ivar Ctypes.structure Ctypes_static.ptr

Changes the value of an instance variable of a class instance.

val ivar_ptr : self:Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes.ptr -> ivar_name:string -> unit Ctypes.ptr

Returns pointer to an ivar in object self

val setter_name_of_ivar : string -> string

Returns the ivar name capitalized, prefixed by "set", suffixed by ":"