Module Runtime.Sel

val register_name : string -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr

Registers a method with the Objective-C runtime system, maps the method name to a selector, and returns the selector value. If the method name has already been registered, this function simply returns the selector.

val get_name : Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr -> string

Returns the name of the method specified by a given selector.

val is_equal : Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr -> bool

Returns a Boolean value that indicates whether two selectors are equal.

val register_typed_name : string -> string -> Runtime__.Type_description.Types.objc_selector Ctypes.structure Ctypes_static.ptr